Externalisation du sélecteur de données enquête
